+++ ModTek/Features/Logging/Appender.cs
@@ -1,0 +1,54 @@
+﻿using System;
+using System.IO;
+using HBS.Logging;
+using ModTek.Util;
+using Object = UnityEngine.Object;
+
+namespace ModTek.Features.Logging
+{
+    internal class Appender : ILogAppender, IDisposable
+    {
+        private readonly StreamWriter writer;
+        private readonly Formatter formatter;
+
+        internal Appender(string path)
+        {
+            writer = new StreamWriter(path) { AutoFlush = true };
+            writer.WriteLine($"{DateTime.Now} ModTek v{VersionTools.LongVersion}");
+            writer.WriteLine(new string('-', 80));
+            writer.WriteLine(VersionInfo.GetFormattedInfo());
+        }
+
+        internal Appender(string path, FormatterSettings settings) : this(path)
+        {
+            formatter = new Formatter(settings);
+        }
+
+        public void OnLogMessage(string logName, LogLevel level, object message, Object context, Exception exception, IStackTrace location)
+        {
+            var logLine = formatter.GetFormattedLogLine(logName, level, message, exception, location);
+            WriteLine(logLine.Line);
+        }
+
+        public void Flush()
+        {
+            // AutoFlush = true
+        }
+
+        public void Dispose()
+        {
+            lock(this)
+            {
+                writer?.Dispose();
+            }
+        }
+
+        internal void WriteLine(string line)
+        {
+            lock(this)
+            {
+                writer.WriteLine(line);
+            }
+        }
+    }
+}