+++ ModTek/Features/Logging/LoggingSettings.cs
@@ -8,9 +8,24 @@ namespace ModTek.Features.Logging
         public string[] PrefixesToIgnore = {};
 
         [JsonProperty]
-        public bool preserveFullLog;
+        public bool PreserveFullLog;
 
         [JsonProperty]
-        public bool skipOriginalLoggers;
+        public bool IgnoreLoggerLogLevel;
+
+        [JsonProperty]
+        public bool SkipOriginalLoggers;
+
+        [JsonProperty]
+        public string[] IgnoreSkipForLoggers = {"ModTek"};
+
+        [JsonProperty]
+        public FormatterSettings ModTekLogFormatting = new()
+        {
+            FormatLine = "{2}{3}" // log level is always LOG and we know its ModTek
+        };
+
+        [JsonProperty]
+        public FormatterSettings BattleTechLogFormatting = new();
     }
 }