+++ ModTek/Features/Logging/RTLogFileType.cs
@@ -1,7 +1,0 @@
-namespace ModTek.Features.Logging
-{
-    internal enum RTLogFileType
-    {
-        Main
-    }
-}